layout/generic/test/test_bug527306.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=527306
-->
<head>
  <title>Test for Bug 527306</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>
<body onload="showText()">
<p><a target="_blank" href="https://bugzilla.mozilla.org/show_bug?id=527306">Mozilla Bug 527306</a>

<p><input type="text" id="currenttext" value="FAIL">

<pre id="test">
<script>
function showText() {
  var t = document.getElementById("currenttext");
  t.style.display = "none";
  t.value = "PASS";
  is(t.value, "PASS", "Value should be set correctly");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>

</body>
</html>
